refactor: move data prepering services to their respective components
